% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{Burden Estimate Sets}
\alias{Burden Estimate Sets}
\title{Burden Estimate Sets}
\arguments{
\item{location}{A montagu location}
}
\description{
Burden estimate sets define a set of results for a modelling group. They
are specific to a touchstone and scenario. Usually, they will be the 
populated version of the burden estimate set template, which a modelling
group can download, and defines the columns and rows for all the 
countries, ages and years that are expected from that group, for that
scenario. The modelling group then overwrites the missing values with
results from their model, and submits the results to Montagu.
}
