% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_groups.R
\name{montagu_add_modelling_group}
\alias{montagu_add_modelling_group}
\title{Add a new modelling group}
\usage{
montagu_add_modelling_group(modelling_group_id,
  modelling_group_description, modelling_group_institution,
  modelling_group_pi, location = NULL)
}
\arguments{
\item{modelling_group_id}{Id for new group, typicallly in form Institution-LeadModeller.}

\item{modelling_group_description}{A text description of the group}

\item{modelling_group_institution}{Name of institution}

\item{modelling_group_pi}{The principal investigator for the group.}

\item{location}{The montagu server to connect to.}
}
\description{
Add a new modelling group
}
