% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{montagu_burden_estimate_set_clear}
\alias{montagu_burden_estimate_set_clear}
\title{Deletes all uploaded rows from an incomplete burden estimate set}
\usage{
montagu_burden_estimate_set_clear(modelling_group_id, touchstone_id,
  scenario_id, burden_estimate_set_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{Modelling group identifier}

\item{touchstone_id}{Touchstone identifier}

\item{scenario_id}{Scenario identifier}

\item{burden_estimate_set_id}{Burden estimate set created by
\code{montagu_burden_estimate_set_crete}}

\item{location}{The montagu server to connect to.}
}
\value{
The id of the burden estimate set
}
\description{
Deletes all uploaded rows from an incomplete burden estimate set
}
