% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{montagu_burden_estimate_set_close}
\alias{montagu_burden_estimate_set_close}
\title{Closes a burden estimate set, marking it as complete.}
\usage{
montagu_burden_estimate_set_close(modelling_group_id, touchstone_id,
  scenario_id, burden_estimate_set_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{Modelling group identifier}

\item{touchstone_id}{Touchstone identifier}

\item{scenario_id}{Scenario identifier}

\item{burden_estimate_set_id}{Burden estimate set created by
\code{montagu_burden_estimate_set_create}}

\item{location}{The montagu server to connect to.}
}
\value{
The id of the burden estimate set
}
\description{
Burden estimate sets define a set of results for a modelling group. They
are specific to a touchstone and scenario. Usually, they will be the
populated version of the burden estimate set template, which a modelling
group can download, and defines the columns and rows for all the
countries, ages and years that are expected from that group, for that
scenario. The modelling group then overwrites the missing values with
results from their model, and submits the results to Montagu.
}
