% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{montagu_burden_estimate_set_create}
\alias{montagu_burden_estimate_set_create}
\title{Create a new burden estimate set}
\usage{
montagu_burden_estimate_set_create(modelling_group_id, touchstone_id,
  scenario_id, type, model_run_parameter_set = NULL, details = NULL,
  location = NULL)
}
\arguments{
\item{modelling_group_id}{Modelling group identifier}

\item{touchstone_id}{Touchstone identifier}

\item{scenario_id}{Scenario identifier}

\item{type}{Can be `central-single-run`, `central-averaged`
or `stochastic`}

\item{model_run_parameter_set}{Identifier for the parameter set}

\item{details}{Optional details string}

\item{location}{The montagu server to connect to.}
}
\value{
The id of the burden estimate set
}
\description{
Create a new burden estimate set
}
