% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{montagu_burden_estimate_set_outcome_data}
\alias{montagu_burden_estimate_set_outcome_data}
\title{Retrieve data for a particular outcoe of a burden estimate set,
aggregated across country and disaggregated by either age or year.}
\usage{
montagu_burden_estimate_set_outcome_data(modelling_group_id, touchstone_id,
  scenario_id, burden_estimate_set_id, outcome_code, group_by = "age",
  location = NULL)
}
\arguments{
\item{modelling_group_id}{Modelling group identifier}

\item{touchstone_id}{Touchstone identifier}

\item{scenario_id}{Scenario identifier}

\item{burden_estimate_set_id}{The integer id of a burden estimate set}

\item{outcome_code}{The name of an outcome, such as 'cases' or 'deaths'.}

\item{group_by}{Set to 'age' (the default) or 'year', to set the}

\item{location}{The montagu server to connect to.}
}
\value{
A data frame with columns age or year (depending on group_by),
}
\description{
Burden estimate sets define a set of results for a modelling group. They
are specific to a touchstone and scenario. Usually, they will be the
populated version of the burden estimate set template, which a modelling
group can download, and defines the columns and rows for all the
countries, ages and years that are expected from that group, for that
scenario. The modelling group then overwrites the missing values with
results from their model, and submits the results to Montagu.
}
