% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{montagu_burden_estimate_set_problems}
\alias{montagu_burden_estimate_set_problems}
\title{Retrieves list of estimate sets for a group, touchstone and scenario.}
\usage{
montagu_burden_estimate_set_problems(modelling_group_id, touchstone_id,
  scenario_id, burden_estimate_set_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{Modelling group identifier}

\item{touchstone_id}{Touchstone identifier}

\item{scenario_id}{Scenario identifier}

\item{burden_estimate_set_id}{Burden estimate set identifier}

\item{location}{The montagu server to connect to.}
}
\value{
A list of any problems with this burden estimate set
}
\description{
Retrieves list of estimate sets for a group, touchstone and scenario.
}
