% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{montagu_burden_estimate_set_request_upload}
\alias{montagu_burden_estimate_set_request_upload}
\title{Requests a signed upload token, that can be used to upload
a file in chunks, and then populate a burden estimate set with the
uploaded file. This can only be done on a central estimate set, not a
stochastic set.}
\usage{
montagu_burden_estimate_set_request_upload(modelling_group_id,
  touchstone_id, scenario_id, burden_estimate_set_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{Modelling group identifier}

\item{touchstone_id}{Touchstone identifier}

\item{scenario_id}{Scenario identifier}

\item{burden_estimate_set_id}{Burden estimate set created by
\code{montagu_burden_estimate_set_crete}}

\item{location}{The montagu server to connect to.}
}
\value{
a token string.
}
\description{
Requests a signed upload token, that can be used to upload
a file in chunks, and then populate a burden estimate set with the
uploaded file. This can only be done on a central estimate set, not a
stochastic set.
}
