% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{montagu_burden_estimate_set_upload}
\alias{montagu_burden_estimate_set_upload}
\title{Deletes all uploaded rows from an incomplete burden estimate set}
\usage{
montagu_burden_estimate_set_upload(modelling_group_id, touchstone_id,
  scenario_id, burden_estimate_set_id, data, lines = 10000L,
  keep_open = FALSE, location = NULL)
}
\arguments{
\item{modelling_group_id}{Modelling group identifier}

\item{touchstone_id}{Touchstone identifier}

\item{scenario_id}{Scenario identifier}

\item{burden_estimate_set_id}{Burden estimate set created by
\code{montagu_burden_estimate_set_crete}}

\item{data}{Data frame containing burden estimates.}

\item{lines}{Number of lines to chunk the files into}

\item{keep_open}{Keep the burden estimate set open after upload?}

\item{location}{The montagu server to connect to.}
}
\value{
The id of the burden estimate set
}
\description{
Burden estimate sets define a set of results for a modelling group. They
are specific to a touchstone and scenario. Usually, they will be the
populated version of the burden estimate set template, which a modelling
group can download, and defines the columns and rows for all the
countries, ages and years that are expected from that group, for that
scenario. The modelling group then overwrites the missing values with
results from their model, and submits the results to Montagu.
}
