% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{montagu_burden_estimate_sets}
\alias{montagu_burden_estimate_sets}
\title{Retrieves list of estimate sets for a group, touchstone and scenario.}
\usage{
montagu_burden_estimate_sets(modelling_group_id, touchstone_id,
  scenario_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{Modelling group identifier}

\item{touchstone_id}{Touchstone identifier}

\item{scenario_id}{Scenario identifier}

\item{location}{The montagu server to connect to.}
}
\value{
A data frame of information about all relevant estimate sets.
}
\description{
Retrieves list of estimate sets for a group, touchstone and scenario.
}
