% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_central_burden_estimate_template}
\alias{montagu_central_burden_estimate_template}
\title{Get central burden estimate template for an expectation}
\usage{
montagu_central_burden_estimate_template(modelling_group_id, touchstone_id,
  expectation_id, location)
}
\arguments{
\item{modelling_group_id}{id of the modelling group.}

\item{touchstone_id}{id of the touchstone (includes version suffix)}

\item{expectation_id}{id of the expectation (integer)}

\item{location}{The montagu server to connect to.}
}
\value{
A data frame with columns disease, year, age, country, and
        country_name with given values, then cohort_size, deaths,
        cases and dalys, all NA.
}
\description{
Get central burden estimate template for an expectation
}
