% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{montagu_coverage_data}
\alias{montagu_coverage_data}
\title{Retrieve coverage data}
\usage{
montagu_coverage_data(modelling_group, touchstone_id, scenario_id,
  format = "long", all_countries = FALSE, location = NULL)
}
\arguments{
\item{modelling_group}{Id of modelling group}

\item{touchstone_id}{Id of touchstone}

\item{scenario_id}{Id of scenario within touchstone}

\item{format}{"wide" or "long" csv format.}

\item{all_countries}{Include all countries, instead of just those
for which burden estimates are required.}

\item{location}{The montagu server to connect to.}
}
\value{
A data frame of metadata about all associated coverage sets
}
\description{
Retrieve coverage data
}
