% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{montagu_coverage_info}
\alias{montagu_coverage_info}
\title{Retrieve metadata for coverage sets associated with a responsibility}
\usage{
montagu_coverage_info(modelling_group, touchstone_id, scenario_id,
  location = NULL)
}
\arguments{
\item{modelling_group}{Id of modelling group}

\item{touchstone_id}{Id of touchstone}

\item{scenario_id}{Id of scenario within touchstone}

\item{location}{The montagu server to connect to.}
}
\value{
A data frame of metadata about all associated coverage sets
}
\description{
This may get a name change pending splitting apart of various
montagu components.
}
