% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_current_estimate_set_info}
\alias{montagu_current_estimate_set_info}
\title{Get information on current estimate set for a scenario.}
\usage{
montagu_current_estimate_set_info(modelling_group_id, touchstone_id,
  scenario_id, location = NULL)
}
\value{
A list of fields about the current estimate set
}
\description{
This may get a name change pending splitting apart of various
montagu components.
}
