% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{montagu_demographics}
\alias{montagu_demographics}
\alias{montagu_demographics_list}
\title{The Montagu API: demographics}
\usage{
montagu_demographics_list(touchstone_id, location = NULL)
}
\arguments{
\item{touchstone_id}{Touchstone identifier}

\item{location}{A montagu location}
}
\description{
This may get a name change pending splitting apart of various
montagu components.
}
