% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{montagu_demographics_list}
\alias{montagu_demographics_list}
\title{List the demographic data available for a given touchstone.}
\usage{
montagu_demographics_list(touchstone_id, location = NULL)
}
\arguments{
\item{touchstone_id}{Touchstone identifier}

\item{location}{The location of the montagu api server.}
}
\value{
A data frame, giving the code and description for each demographic
statistic, whether the data are stratified by gender, and a code indicating
the source of the demographic data, which will change if there are future
updates to any of the demographic data fields.
}
\description{
Montagu provides standardised demographic data for the groups to use in
their models. The source of the data is UNWPP, although various procedures
are applied for convenience, including work on some smaller countries,
people above the age of 80 in certain time periods, and extrapolating
both cohorts backwards in time to their origins before 1950. Internal
documents in the reporting portal describe the methods and motivations for
these extensions.
}
