% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.R
\name{montagu_disease}
\alias{montagu_disease}
\title{Retrieve disease with given id.}
\usage{
montagu_disease(disease_id, location = NULL)
}
\arguments{
\item{disease_id}{The disease id.}

\item{location}{the montagu server to connect to.}
}
\value{
a list of an id and a disease name
}
\description{
Diseases have an internal id and a name in Montagu, which are generally,
but not always the same. Occasionally, it is necessary to provide the
disease id to Montagu functions.
}
