% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.R
\name{montagu_disease_by_id}
\alias{montagu_disease_by_id}
\title{Retrieve disease with given id.}
\usage{
montagu_disease_by_id(disease_id, location = NULL)
}
\arguments{
\item{disease_id}{The disease id.}

\item{location}{The montagu server to connect to.}
}
\description{
Retrieve disease with given id.
}
