% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.R
\name{montagu_disease_list}
\alias{montagu_disease_list}
\title{Retrieve list of all disease ids and names.}
\usage{
montagu_disease_list(location = NULL)
}
\arguments{
\item{location}{The montagu server to connect to.}
}
\description{
This may get a name change pending splitting apart of various
montagu components.
}
