% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.R
\name{montagu_diseases}
\alias{montagu_diseases}
\title{Retrieve list of all disease ids and names.}
\usage{
montagu_diseases(location = NULL)
}
\arguments{
\item{location}{the montagu server to connect to.}
}
\value{
data frame containing ids and names of diseases
}
\description{
Diseases have an internal id and a name in Montagu, which are generally,
but not always the same. Occasionally, it is necessary to provide the
disease id to Montagu functions.
}
