% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_expectation}
\alias{montagu_expectation}
\title{Get information about an expectation}
\usage{
montagu_expectation(modelling_group_id, touchstone_id, expectation_id,
  location = NULL)
}
\arguments{
\item{modelling_group_id}{Id of the modelling group.}

\item{touchstone_id}{Id of the touchstone (includes version suffix)}

\item{expectation_id}{Integer ID of the expectation}

\item{location}{The montagu server to connect to.}
}
\value{
A list of data about the expectation
}
\description{
This may get a name change pending splitting apart of various
montagu components.
}
