% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_expectation_countries}
\alias{montagu_expectation_countries}
\title{Get country list for an expectation}
\usage{
montagu_expectation_countries(modelling_group_id, touchstone_id,
  expectation_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{id of the modelling group.}

\item{touchstone_id}{id of the touchstone (includes version suffix)}

\item{expectation_id}{Integer ID of the expectation}

\item{location}{The montagu server to connect to.}
}
\value{
A data frame of country id and name
}
\description{
Get country list for an expectation
}
