% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_expectation_outcomes}
\alias{montagu_expectation_outcomes}
\title{Get expected outcomes}
\usage{
montagu_expectation_outcomes(modelling_group_id, touchstone_id,
  expectation_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{id of the modelling group. If omitted, or null, 
then all responsibility sets associated with the touchstone are returned.}

\item{touchstone_id}{id of the touchstone (including version)}

\item{expectation_id}{id of the expectation (integer)}
}
\value{
A vector of outcome names
}
\description{
The expectations, for a modelling group for a particular touchstone, indicate the
range of chronological years, ages and countries for which burden estimates
are expected, and what burden outcomes are required.
}
