% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_expectations}
\alias{montagu_expectations}
\title{Get expectations for a modelling group's touchstone}
\usage{
montagu_expectations(modelling_group_id, touchstone_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{id of the modelling group.}

\item{touchstone_id}{id of the touchstone (includes version suffix)}

\item{location}{The montagu server to connect to.}
}
\value{
A data frame of information about the expectation
}
\description{
Get expectations for a modelling group's touchstone
}
