% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{montagu_model}
\alias{montagu_model}
\title{Retrieve model with given id.}
\usage{
montagu_model(model_id, location = NULL)
}
\arguments{
\item{model_id}{The model id.}

\item{location}{The montagu server to connect to.}
}
\value{
List containing information about one model
}
\description{
Models and modelling groups have text ids in Montagu. Here we can look up
the relationships between groups and models, with description and citation
information.
}
