% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{montagu_model_by_id}
\alias{montagu_model_by_id}
\title{Retrieve model with given id.}
\usage{
montagu_model_by_id(model_id, location = NULL)
}
\arguments{
\item{model_id}{The model id.}

\item{location}{The montagu server to connect to.}
}
\description{
Retrieve model with given id.
}
