% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_groups.R
\name{montagu_model_list_by_modelling_group_id}
\alias{montagu_model_list_by_modelling_group_id}
\title{Retrieve list of models for a given modelling_group_id.}
\usage{
montagu_model_list_by_modelling_group_id(modelling_group_id,
  location = NULL)
}
\arguments{
\item{modelling_group_id}{The id of the modelling group}

\item{location}{The montagu server to connect to.}
}
\description{
Retrieve list of models for a given modelling_group_id.
}
