% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{montagu_model_run_parameter_set_data}
\alias{montagu_model_run_parameter_set_data}
\title{Retrieve info about one model run parameter sets for a group and touchstone}
\usage{
montagu_model_run_parameter_set_data(modelling_group_id, touchstone_id,
  model_run_parameter_set_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{The id of the modelling group}

\item{touchstone_id}{The id of the touchstone}

\item{model_run_parameter_set_id}{The id of the model_run_parameter_set.}

\item{location}{The montagu server to connect to.}
}
\value{
a csv of parameter values for each run_id.
}
\description{
A model run parameter set is a list of parameters that is required
for a set of stochastic model runs. In a stochastic ensemble, modelling
groups are requested to perform a number of model runs. Each model run
executes with a unique set of parameters, and for a given run, all scenarios
are run with the same parameters, so that we can calculate impact between
them. The spread of parameters across the different runs should capture
the range of sensible behaviour of the model. The model run parameter set then
contains as many rows as there are model runs. Each row must contain a run_id,
and the value for each parameter that is varied.
}
\details{
Adding, and querying existing model_run_parameter_sets is supported, and
when creating a stochastic burden estimate set, the id of the associated
model_run_parameter_set is required as a parameter.
}
