% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{montagu_model_run_parameter_set_info}
\alias{montagu_model_run_parameter_set_info}
\title{Retrieve info about one model run parameter sets for a group and touchstone}
\usage{
montagu_model_run_parameter_set_info(modelling_group_id, touchstone_id,
  model_run_parameter_set_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{The id of the modelling group}

\item{touchstone_id}{The id of the touchstone}

\item{model_run_parameter_set_id}{The id of the model_run_parameter_set.}

\item{location}{The montagu server to connect to.}
}
\value{
a list of info about one  model run parameter set.
}
\description{
This may get a name change pending splitting apart of various
montagu components.
}
