% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{montagu_model_run_parameter_set_upload}
\alias{montagu_model_run_parameter_set_upload}
\title{Upload a mode_run_parameter_set to Montagu.}
\usage{
montagu_model_run_parameter_set_upload(modelling_group_id, touchstone_id,
  disease_id, params, location = NULL)
}
\arguments{
\item{modelling_group_id}{The id of the modelling group}

\item{touchstone_id}{The id of the touchstone}

\item{disease_id}{The id of the disease associated with this model}

\item{location}{The montagu server to connect to.}

\item{data}{a data frame with a column `run_id`, and other columns for each parameter that varies by run.}
}
\value{
the id of the newly created model_run_parameter_set
}
\description{
This may get a name change pending splitting apart of various
montagu components.
}
