% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{montagu_model_run_parameter_sets}
\alias{montagu_model_run_parameter_sets}
\title{Retrieve list of model run parameter sets for a group and touchstone}
\usage{
montagu_model_run_parameter_sets(modelling_group_id, touchstone_id,
  location = NULL)
}
\arguments{
\item{modelling_group_id}{The id of the modelling group}

\item{touchstone_id}{The id of the touchstone}

\item{location}{The montagu server to connect to.}
}
\value{
a data frame of info about the model run parameter sets.
}
\description{
This may get a name change pending splitting apart of various
montagu components.
}
