% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_groups.R
\name{montagu_modelling_group}
\alias{montagu_modelling_group}
\title{Retrieve modelling_group with given modelling_group_id.}
\usage{
montagu_modelling_group(modelling_group_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{The id of the modelling group}

\item{location}{The montagu server to connect to.}
}
\value{
A list of id and description for the given group id.
}
\description{
Modelling group ids generally consist of an identifier for the
institution where the group is based, a hyphen, and the principal
investigor for that group. Groups will be informed of their internal 
id when joining the consortium, but for convenience, we provide
functions to look up the modelling group id, and associated models.
}
