% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_groups.R
\name{montagu_modelling_group_by_id}
\alias{montagu_modelling_group_by_id}
\title{Retrieve modelling_group with given modelling_group_id.}
\usage{
montagu_modelling_group_by_id(modelling_group_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{The id of the modelling group}

\item{location}{The montagu server to connect to.}
}
\description{
Retrieve modelling_group with given modelling_group_id.
}
