% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_groups.R
\name{montagu_modelling_group_list}
\alias{montagu_modelling_group_list}
\title{Retrieve list of all modelling group ids and names.}
\usage{
montagu_modelling_group_list(location = NULL)
}
\arguments{
\item{location}{The montagu server to connect to.}
}
\description{
This may get a name change pending splitting apart of various
montagu components.
}
