% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_groups.R
\name{montagu_modelling_group_members}
\alias{montagu_modelling_group_members}
\title{Retrieve list of members for a given modelling_group_id.}
\usage{
montagu_modelling_group_members(modelling_group_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{The id of the modelling group}

\item{location}{The montagu server to connect to.}
}
\description{
Retrieve list of members for a given modelling_group_id.
}
