% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{montagu_models_list}
\alias{montagu_models_list}
\title{Retrieve list of all model ids, names, citations and groups.}
\usage{
montagu_models_list(location = NULL)
}
\arguments{
\item{location}{The montagu server to connect to.}
}
\description{
This may get a name change pending splitting apart of various
montagu components.
}
