% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly.R
\name{montagu_orderly_remote}
\alias{montagu_orderly_remote}
\title{orderly remote driver}
\usage{
montagu_orderly_remote(name, host, port = 443, basic = FALSE,
  github = FALSE, username = NULL, password = NULL)
}
\arguments{
\item{name}{Name for the remote}

\item{host}{Hostname}

\item{port}{Port}

\item{basic}{Logical, indicating if this remote uses basic authentication}

\item{github}{Logical, indicating if the server is secured with
OrderlyWeb's GitHub flow.  If provided then the \code{password}
field must contain the github token, or the name of an
environment variable that represents the github token (with a
leading \code{$}, for example \code{$ORDERLY_GITHUB_TOKEN}).}

\item{username}{Username}

\item{password}{Password}
}
\description{
An \code{orderly} remote driver for montagu
}
