% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly.R
\name{montagu_orderlyweb_remote}
\alias{montagu_orderlyweb_remote}
\title{orderly remote driver}
\usage{
montagu_orderlyweb_remote(name, host, port, ...)
}
\arguments{
\item{name}{Name for the remote, shared with the
\code{\link{montagu_server}} name.}

\item{host}{Hostname, shared with the
\code{\link{montagu_server}} host.}

\item{port}{Port, shared with the
\code{\link{montagu_server}} port.}

\item{...}{Additional arguments passed through to
\code{\link{montagu_server}}}
}
\description{
An \code{orderlyweb} remote driver for montagu
}
