% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_groups.R
\name{montagu_remove_user_from_modelling_group}
\alias{montagu_remove_user_from_modelling_group}
\title{Remove user from a modelling group}
\usage{
montagu_remove_user_from_modelling_group(modelling_group_id, username,
  location = NULL)
}
\arguments{
\item{modelling_group_id}{The id of the modelling group}

\item{username}{User name of the user to remove}

\item{location}{The montagu server to connect to.}
}
\description{
Remove user from a modelling group
}
