% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{montagu_reports}
\alias{montagu_reports}
\alias{montagu_reports_list}
\alias{montagu_reports_report_versions}
\alias{montagu_reports_report_metadata}
\alias{montagu_reports_report_download}
\alias{montagu_reports_report_artefact_list}
\alias{montagu_reports_report_artefact_get}
\alias{montagu_reports_report_resource_list}
\alias{montagu_reports_report_resource_get}
\alias{montagu_reports_report_data_list}
\alias{montagu_reports_report_data_get}
\alias{montagu_reports_report_sessioninfo}
\alias{montagu_reports_data}
\alias{montagu_reports_run}
\alias{montagu_reports_wait}
\alias{montagu_reports_status}
\alias{montagu_reports_publish}
\alias{montagu_reports_rebuild}
\alias{montagu_reports_git_status}
\alias{montagu_reports_git_pull}
\alias{montagu_reports_git_fetch}
\title{The Montagu Reporting API}
\usage{
montagu_reports_list(location = NULL)

montagu_reports_report_versions(name, error_if_missing = TRUE,
  location = NULL)

montagu_reports_report_metadata(name, version, location = NULL)

montagu_reports_report_download(name, version, dest = tempfile(),
  progress = TRUE, location = NULL)

montagu_reports_report_artefact_list(name, version, location = NULL)

montagu_reports_report_artefact_get(name, version, filename, dest = NULL,
  progress = TRUE, location = NULL)

montagu_reports_report_resource_list(name, version, location = NULL)

montagu_reports_report_resource_get(name, version, filename, dest = NULL,
  progress = TRUE, location = NULL)

montagu_reports_report_data_list(name, version, location = NULL)

montagu_reports_report_data_get(name, version, hash, dest = NULL,
  progress = TRUE, csv = FALSE, location = NULL)

montagu_reports_report_sessioninfo(name, version, location = NULL)

montagu_reports_data(hash, csv = FALSE, dest = NULL, progress = TRUE,
  location = NULL)

montagu_reports_run(name, parameters = NULL, ref = NULL,
  update = TRUE, timeout = NULL, wait = Inf, poll = 0.5,
  open = FALSE, stop_on_error = FALSE, stop_on_timeout = TRUE,
  progress = TRUE, location = NULL, output = TRUE)

montagu_reports_wait(obj, timeout = 3600, poll = 0.5, open = FALSE,
  stop_on_error = FALSE, stop_on_timeout = TRUE, progress = TRUE,
  output = TRUE, wait = TRUE)

montagu_reports_status(key, output = FALSE, location = NULL)

montagu_reports_publish(name, version, value = NULL, location = NULL)

montagu_reports_rebuild(location = NULL)

montagu_reports_git_status(location = NULL)

montagu_reports_git_pull(location = NULL)

montagu_reports_git_fetch(location = NULL)
}
\arguments{
\item{location}{A montagu location}

\item{name}{A report name}

\item{error_if_missing}{Throw an error if the report is missing -
this is often not desired.}

\item{version}{The report version identifier (YYYYMMDD-HHMMSS-xxxxxxxx)}

\item{dest}{Destination to download to}

\item{progress}{Print a progress bar}

\item{filename}{Name of artefact to download}

\item{hash}{Hash of data to get}

\item{csv}{Logical, indicating if csv (rather than rds) should be returned}

\item{parameters}{List of parameters to run report with}

\item{ref}{Git reference}

\item{update}{I can't remember?}

\item{timeout}{Time to give up on running report; this is set on
the \emph{server}.  After \code{timeout} seconds, the server
will kill the running report.}

\item{wait}{Integer value indicating time to wait on the client
side for the report to run.  If \code{0} (or less) we do not
wait.  If you provide a timeout you should make wait at least
the timeout value.}

\item{poll}{Time to poll for update}

\item{open}{Open the report in a browser on completion?}

\item{stop_on_error}{Throw an error if the report fails?}

\item{stop_on_timeout}{Throw an error if report is not completed in time?}

\item{output}{Show output from running the report.  This is a work
in progress.  This has an effect only when \code{progress} is
\code{TRUE}.}

\item{obj}{A \code{montagu_reports_instance} object, as created by
\code{montagu_reports_run}.}

\item{key}{Key to a running report (adjective_animal)}

\item{value}{Value to publish report as (boolean)}
}
\description{
This may get a name change pending splitting apart of various
montagu components.
}
