% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{montagu_server}
\alias{montagu_server}
\title{Create a handle to a montagu server}
\usage{
montagu_server(name, hostname, port = 443, basic = FALSE, username = NULL,
  password = NULL, verbose = FALSE, global = TRUE, overwrite = FALSE)
}
\arguments{
\item{name}{The name to call this server.  This appears in a few
places as a "friendly" name.  If \code{global = TRUE} then this
will be available from the package using this name.}

\item{hostname}{The hostname of the server}

\item{port}{The port the server is running on (default is 443
which is the standard https port)}

\item{basic}{Logical, indicating if the server is secured with
http basic auth only.}

\item{username}{The username.  If not given, then on
authorisation, the user will be prompted.  Falls back on the R
global options \code{montagu.<name>.username} and
\code{montagu.username} in turn.}

\item{password}{The password.  If not given, then on
authorisation, the user will be prompted.  Falls back on the R
global options \code{montagu.<name>.password} and
\code{montagu.password} in turn.}

\item{verbose}{Logical, indicating if verbose http communication
should be used.  This is for debugging only.}

\item{global}{Logical, indicating if the server should be stored
in the global set.  If true then you can pass \code{location =
name} to most other functions in the package.}

\item{overwrite}{Logical, and meaningful only if \code{global =
TRUE}, indicating if a global configuration should be replaced
if it exists.}
}
\value{
Invisibly, a \code{montagu_server} object.
}
\description{
Create a handle to a montagu server
}
