% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{montagu_server_global_default_set}
\alias{montagu_server_global_default_set}
\title{Set a global default montagu server}
\usage{
montagu_server_global_default_set(location)
}
\arguments{
\item{location}{A server location}
}
\description{
Set a global default montagu server
}
