% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_stochastic_burden_estimate_template}
\alias{montagu_stochastic_burden_estimate_template}
\title{Get stochastic burden estimate template for an expectation}
\usage{
montagu_stochastic_burden_estimate_template(modelling_group_id,
  touchstone_id, expectation_id, location)
}
\arguments{
\item{modelling_group_id}{id of the modelling group.}

\item{touchstone_id}{id of the touchstone (includes version suffix)}

\item{expectation_id}{id of the expectation (integer)}

\item{location}{The montagu server to connect to.}
}
\value{
A data frame with columns disease (given), run_id (NA), year, age,
        country, and country_name (all given), and finally
        cohort_size, deaths, cases and dalys, (NA).
}
\description{
Get stochastic burden estimate template for an expectation
}
