% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_touchstones}
\alias{montagu_touchstones}
\title{Retrieve touchstones a modelling group is responsible for.}
\usage{
montagu_touchstones(modelling_group_id = NULL, location = NULL)
}
\arguments{
\item{modelling_group_id}{id of the modelling group. If omitted or null, 
then all touchstones are returned.}

\item{location}{The montagu server to connect to.}
}
\value{
Data frame of touchstone name, description and comment.
}
\description{
Touchstones are created by the VIMC management. The touchstone id associates
a particular call for burden estimates, with the input data (coverage and
demography) required to produce those estimates, and with the estimates that
the modelling groups provide. Usually,
}
