% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_touchstones_for_scenario}
\alias{montagu_touchstones_for_scenario}
\title{Get touchstones that include this scenario}
\usage{
montagu_touchstones_for_scenario(modelling_group_id, touchstone_id,
  scenario_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{id of the modelling group. If omitted, or null, 
then all responsibility sets associated with the touchstone are returned.}

\item{touchstone_id}{id of the touchstone (including version)}

\item{scenario_id}{id of the scenario}

\item{location}{The montagu server to connect to.}
}
\value{
vector of touchstone ids
}
\description{
Given a particular touchstone and scenario, we can loookup all the other
touchstones that also contain this scenario.
}
