% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{montagu_burden_estimate_set_data}
\alias{montagu_burden_estimate_set_data}
\alias{montagu_burden_estimate_set_clear}
\alias{montagu_burden_estimate_set_upload}
\title{Retrieves the data for a specific burden estimate set.}
\usage{
montagu_burden_estimate_set_data(modelling_group_id, touchstone_id,
  scenario_id, burden_estimate_set_id, location = NULL)

montagu_burden_estimate_set_clear(modelling_group_id, touchstone_id,
  scenario_id, burden_estimate_set_id, location = NULL)

montagu_burden_estimate_set_upload(modelling_group_id, touchstone_id,
  scenario_id, burden_estimate_set_id, data, lines = 10000L,
  keep_open = FALSE, location = NULL)
}
\arguments{
\item{modelling_group_id}{Modelling group identifier}

\item{touchstone_id}{Touchstone identifier}

\item{scenario_id}{Scenario identifier}

\item{burden_estimate_set_id}{Burden estimate set created by
\code{montagu_burden_estimate_set_crete}}

\item{location}{The montagu server to connect to.}

\item{data}{Data frame containing burden estimates.}

\item{lines}{Number of lines to chunk the files into}

\item{keep_open}{Keep the burden estimate set open after upload?}
}
\value{
A list of information about a specific estimate set.
}
\description{
Retrieves the data for a specific burden estimate set.

Deletes all uploaded rows from an incomplete burden estimate set
}
