% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{montagu_api}
\alias{montagu_api}
\alias{montagu_burden_estimate_sets}
\alias{montagu_burden_estimate_set_info}
\alias{montagu_burden_estimate_set_data}
\alias{montagu_burden_estimate_set_problems}
\alias{montagu_burden_estimate_set_create}
\alias{montagu_burden_estimate_set_upload}
\alias{montagu_burden_estimate_set_clear}
\alias{montagu_touchstones_list}
\title{The Montagu API}
\usage{
montagu_burden_estimate_sets(modelling_group_id, touchstone_id,
  scenario_id, location = NULL)

montagu_burden_estimate_set_info(modelling_group_id, touchstone_id,
  scenario_id, burden_estimate_set_id, location = NULL)

montagu_burden_estimate_set_data(modelling_group_id, touchstone_id,
  scenario_id, burden_estimate_set_id, location = NULL)

montagu_burden_estimate_set_problems(modelling_group_id, touchstone_id,
  scenario_id, burden_estimate_set_id, location = NULL)

montagu_burden_estimate_set_create(modelling_group_id, touchstone_id,
  scenario_id, type, model_run_parameter_set, details = NULL,
  location = NULL)

montagu_burden_estimate_set_upload(modelling_group_id, touchstone_id,
  scenario_id, burden_estimate_set_id, filename, lines = 10000L,
  keep_open = FALSE, location = NULL)

montagu_burden_estimate_set_clear(modelling_group_id, touchstone_id,
  scenario_id, burden_estimate_set_id, location = NULL)

montagu_touchstones_list(location = NULL)
}
\arguments{
\item{modelling_group_id}{Modelling group identifier}

\item{touchstone_id}{Touchstone identifier}

\item{scenario_id}{Scenario identifier}

\item{location}{A montagu location}

\item{burden_estimate_set_id}{Burden estimate set identifier}

\item{type}{Either "stochastic" or "central" (I think)}

\item{model_run_parameter_set}{Identifier for the parameter set}

\item{details}{Optional details string}

\item{filename}{Filename to upload}

\item{lines}{Number of lines to chunk the files into}

\item{keep_open}{Keep the burden estimate set open after upload?}

\item{location}{The montagu server to connect to.}

\item{modelling_group_id}{Modelling group identifier}

\item{touchstone_id}{Touchstone identifier}

\item{scenario_id}{Scenario identifier}

\item{location}{The montagu server to connect to.}

\item{burden_estimate_set_id}{Burden estimate set created by
\code{montagu_burden_estimate_set_crete}}
}
\value{
A data frame of information about all relevant estimate sets.

A list of information about a specific estimate set.

A list of information about a specific estimate set.

A list of any problems with this burden estimate set
}
\description{
This may get a name change pending splitting apart of various
montagu components.
}
