% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{montagu_burden_estimate_set_close}
\alias{montagu_burden_estimate_set_close}
\title{Closes a burden estimate set, marking it as complete.}
\usage{
montagu_burden_estimate_set_close(modelling_group_id, touchstone_id,
  scenario_id, burden_estimate_set_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{Modelling group identifier}

\item{touchstone_id}{Touchstone identifier}

\item{scenario_id}{Scenario identifier}

\item{burden_estimate_set_id}{Burden estimate set created by
\code{montagu_burden_estimate_set_create}}

\item{location}{The montagu server to connect to.}
}
\value{
A list of information about a specific estimate set.
}
\description{
Closes a burden estimate set, marking it as complete.
}
