% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{montagu_burden_estimate_set_info}
\alias{montagu_burden_estimate_set_info}
\title{Retrieves information about a specific burden estimate set.}
\usage{
montagu_burden_estimate_set_info(modelling_group_id, touchstone_id,
  scenario_id, burden_estimate_set_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{Modelling group identifier}

\item{touchstone_id}{Touchstone identifier}

\item{scenario_id}{Scenario identifier}

\item{burden_estimate_set_id}{The integer id of a burden estimate set}

\item{location}{The montagu server to connect to.}
}
\value{
A list of information about a specific estimate set.
}
\description{
Burden estimate sets define a set of results for a modelling group. They
are specific to a touchstone and scenario. Usually, they will be the
populated version of the burden estimate set template, which a modelling
group can download, and defines the columns and rows for all the
countries, ages and years that are expected from that group, for that
scenario. The modelling group then overwrites the missing values with
results from their model, and submits the results to Montagu.
}
