% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_central_burden_estimate_template}
\alias{montagu_central_burden_estimate_template}
\title{Get central burden estimate template for expectation}
\usage{
montagu_central_burden_estimate_template(modelling_group_id, touchstone_id,
  expectation_id, location)
}
\arguments{
\item{modelling_group_id}{Id of the modelling group.}

\item{touchstone_id}{Id of the touchstone (includes version suffix)}

\item{expectation_id}{Integer ID of the expectation}

\item{location}{The montagu server to connect to.}
}
\value{
A data frame with columns disease, year, age, country, and
        country_name with given values, then cohort_size, deaths,
        cases and dalys, all NA.
}
\description{
This may get a name change pending splitting apart of various
montagu components.
}
