% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{montagu_coverage_info}
\alias{montagu_coverage_info}
\title{Retrieve metadata for coverage sets for a particularly group, touchstone and scenario.}
\usage{
montagu_coverage_info(modelling_group, touchstone_id, scenario_id,
  location = NULL)
}
\arguments{
\item{modelling_group}{id of modelling group}

\item{touchstone_id}{id of touchstone (including version)}

\item{scenario_id}{id of scenario within touchstone}

\item{location}{The montagu server to connect to.}
}
\value{
A data frame of metadata about all associated coverage sets
}
\description{
The coverage data describes for each country, over time and age, the
efforts to vaccinate people against each disease. Routine activities involve
the regular vaccination that is in place, whereas campaigns are specifically
targeted towards a certain disease or people group for a certain time period.
}
