% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_current_estimate_set_info}
\alias{montagu_current_estimate_set_info}
\title{Get information on current estimate set for a scenario.}
\usage{
montagu_current_estimate_set_info(modelling_group_id, touchstone_id,
  scenario_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{id of the modelling group.}

\item{touchstone_id}{id of the touchstone (includes version suffix)}

\item{scenario_id}{id of the scenario}

\item{location}{The montagu server to connect to.}
}
\value{
A list of fields about the current estimate set
}
\description{
Get information on current estimate set for a scenario.
}
