% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_current_estimate_set_problems}
\alias{montagu_current_estimate_set_problems}
\title{Get information on current estimate set for a scenario.}
\usage{
montagu_current_estimate_set_problems(modelling_group_id, touchstone_id,
  scenario_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{Id of the modelling group.}

\item{touchstone_id}{Id of the touchstone (includes version suffix)}

\item{scenario_id}{Id of the scenario}

\item{location}{The montagu server to connect to.}
}
\value{
A list of problems.
}
\description{
This may get a name change pending splitting apart of various
montagu components.
}
