% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{montagu_demographic_data}
\alias{montagu_demographic_data}
\title{Download demographic data for a given touchstone.}
\usage{
montagu_demographic_data(type_code, touchstone_id, gender_code = NULL,
  wide = FALSE, source_code = NULL, location = NULL)
}
\arguments{
\item{type_code}{The demographic type code (something like
\code{cbr} for \code{Fertility: Crude birth rate (CBR)}.  Get
the possible values from \code{montagu_demographics_list}}

\item{gender_code}{One of "male", "female" or "both" (if omitted
montagu will return "both").  Not applicable to all statistics.}

\item{wide}{Logical, if \code{TRUE} returns data in wide format
rather than long.}

\item{source_code}{Optional source code, may be useful if one
\code{type_code} can come from two sources within a touchstone.}
}
\description{
Download demographic data for a given touchstone.
}
