% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseases.R
\name{montagu_diseases}
\alias{montagu_diseases}
\title{Retrieve list of all disease ids and names.}
\usage{
montagu_diseases(location = NULL)
}
\arguments{
\item{location}{the montagu server to connect to.}
}
\description{
Retrieve list of all disease ids and names.
}
