% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_expectation}
\alias{montagu_expectation}
\title{Get information about an expectation}
\usage{
montagu_expectation(modelling_group_id, touchstone_id, expectation_id,
  location = NULL)
}
\arguments{
\item{modelling_group_id}{id of the modelling group.}

\item{touchstone_id}{id of the touchstone (includes version suffix)}

\item{expectation_id}{id of the expectation (integer)}

\item{location}{The montagu server to connect to.}
}
\value{
list of data about the expectation
}
\description{
Get information about an expectation
}
