% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_expectation_applicable_scenarios}
\alias{montagu_expectation_applicable_scenarios}
\title{Get applicable scenarios for an expectation}
\usage{
montagu_expectation_applicable_scenarios(modelling_group_id, touchstone_id,
  expectation_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{Id of the modelling group.}

\item{touchstone_id}{Id of the touchstone (includes version suffix)}

\item{expectation_id}{Integer ID of the expectation}

\item{location}{The montagu server to connect to.}
}
\value{
A vector of scenario names
}
\description{
This may get a name change pending splitting apart of various
montagu components.
}
