% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{montagu_model}
\alias{montagu_model}
\title{Retrieve model with given id.}
\usage{
montagu_model(model_id, location = NULL)
}
\arguments{
\item{model_id}{The model id.}
}
\description{
Retrieve model with given id.
}
