% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{montagu_model_run_parameter_set_upload}
\alias{montagu_model_run_parameter_set_upload}
\title{Upload a model_run_parameter_set to Montagu.}
\usage{
montagu_model_run_parameter_set_upload(modelling_group_id, touchstone_id,
  disease_id, data, location = NULL)
}
\arguments{
\item{modelling_group_id}{The id of the modelling group}

\item{touchstone_id}{The id of the touchstone}

\item{disease_id}{The id of the disease associated with this model}

\item{data}{a data frame with a column `run_id`, and other columns for each 
parameter that will be varied for each run.}

\item{location}{The montagu server to connect to.}
}
\value{
the id of the newly created model_run_parameter_set
}
\description{
A model run parameter set is a list of parameters that is required
for a set of stochastic model runs. In a stochastic ensemble, modelling
groups are requested to perform a number of model runs. Each model run
executes with a unique set of parameters, and for a given run, all scenarios
are run with the same parameters, so that we can calculate impact between
them. The spread of parameters across the different runs should capture
the range of sensible behaviour of the model. The model run parameter set then
contains as many rows as there are model runs. Each row must contain a run_id,
and the value for each parameter that is varied.
}
\details{
Adding, and querying existing model_run_parameter_sets is supported, and
when creating a stochastic burden estimate set, the id of the associated
model_run_parameter_set is required as a parameter.
}
