% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{montagu_models}
\alias{montagu_models}
\title{Retrieve list of all model ids, names, citations and groups.}
\usage{
montagu_models(location = NULL)
}
\arguments{
\item{location}{The montagu server to connect to.}
}
\value{
Data frame containing information about models.
}
\description{
Models and modelling groups have text ids in Montagu. Here we can look up
the relationships between groups and models, with description and citation
information.
}
