% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly.R
\name{montagu_orderly_remote}
\alias{montagu_orderly_remote}
\title{orderly remote driver}
\usage{
montagu_orderly_remote(name, host, port = 443, basic = FALSE,
  username = NULL, password = NULL)
}
\arguments{
\item{name}{Name for the remote}

\item{host}{Hostname}

\item{port}{Port}

\item{basic}{Logical, indicating if this remote uses basic authentication}

\item{username}{Username}

\item{password}{Password}
}
\description{
An \code{orderly} remote driver for montagu
}
