% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_scenario_problems}
\alias{montagu_scenario_problems}
\title{Retrieve a list of any problems with a scenario.}
\usage{
montagu_scenario_problems(modelling_group_id, touchstone_id, scenario_id,
  location = NULL)
}
\arguments{
\item{modelling_group_id}{id of the modelling group. If omitted, or null, 
then all responsibility sets associated with the touchstone are returned.}

\item{touchstone_id}{id of the touchstone (including version)}

\item{scenario_id}{id of the scenario}
}
\value{
A list of problem text.
}
\description{
A scenario describes the vaccination conditions for a particular run of a
model. Typical examples include a scenario where there is no vaccination,
a scenario where there is routine (background) vaccination, or a scenario
in which there are targetted campaigns to vaccinate particularly ages. 
Depending on disease, modelling groups may be asked to model various
scenarios, for a particular touchstone.
}
