% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_scenario_status}
\alias{montagu_scenario_status}
\title{Retrieve current status of a groups' scenario.}
\usage{
montagu_scenario_status(modelling_group_id, touchstone_id, scenario_id,
  location = NULL)
}
\arguments{
\item{modelling_group_id}{id of the modelling group. If omitted, or null, 
then all responsibility sets associated with the touchstone are returned.}

\item{touchstone_id}{id of the touchstone (including version)}

\item{scenario_id}{id of the scenario}

\item{location}{The montagu server to connect to.}
}
\value{
"invalid", "complete", "valid", or "empty"
}
\description{
A scenario describes the vaccination conditions for a particular run of a
model. Typical examples include a scenario where there is no vaccination,
a scenario where there is routine (background) vaccination, or a scenario
in which there are targetted campaigns to vaccinate particularly ages.
Depending on disease, modelling groups may be asked to model various
scenarios, for a particular touchstone.
}
