% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_scenarios}
\alias{montagu_scenarios}
\title{Retrieve information about a scenario}
\usage{
montagu_scenarios(modelling_group_id, touchstone_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{id of the modelling group.}

\item{touchstone_id}{id of the touchstone (including version)}

\item{location}{The montagu server to connect to.}
}
\value{
Data frame of scenario_id, description and disease.
}
\description{
A scenario describes the vaccination conditions for a particular run of a
model. Typical examples include a scenario where there is no vaccination,
a scenario where there is routine (background) vaccination, or a scenario
in which there are targetted campaigns to vaccinate particularly ages. 
Depending on disease, modelling groups may be asked to model various
scenarios, for a particular touchstone.
}
