% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_scenarios}
\alias{montagu_scenarios}
\title{Retrieve information about a scenario}
\usage{
montagu_scenarios(modelling_group_id, touchstone_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{id of the modelling group.}

\item{touchstone_id}{id of the touchstone (includes version suffix)}

\item{location}{The montagu server to connect to.}
}
\value{
Data frame of scenario_id, description and disease.
}
\description{
Retrieve information about a scenario
}
