% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_touchstone_versions}
\alias{montagu_touchstone_versions}
\title{Retrieve list of all versions of a given touchstone.}
\usage{
montagu_touchstone_versions(modelling_group_id, touchstone_name,
  location = NULL)
}
\arguments{
\item{modelling_group_id}{Id of the modelling group.}

\item{touchstone_name}{Name of the touchstone (ie, no version suffix)}

\item{location}{The montagu server to connect to.}
}
\value{
Data frame of touchstine id, name, version, description and status
}
\description{
Retrieve list of all versions of a given touchstone.
}
