% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_touchstone_versions}
\alias{montagu_touchstone_versions}
\title{Retrieve list of all versions of a given touchstone.}
\usage{
montagu_touchstone_versions(modelling_group_id, touchstone_name = NULL,
  require_open = FALSE, location = NULL)
}
\arguments{
\item{modelling_group_id}{id of the modelling group.}

\item{touchstone_name}{Optional base name of the touchstone to filter. (ie, no version suffix)}

\item{require_open}{Only include open touchstones if set to true.}

\item{location}{The montagu server to connect to.}
}
\value{
Data frame of touchstone id, name, version, description and status
}
\description{
Touchstones are versioned; errata in the input or coverage data may be
addressed with a new version of the existing touchstone. When interacting
with Montagu, a touchstone_id will consist of a basename, and a version.
}
