% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsibilities.R
\name{montagu_touchstones}
\alias{montagu_touchstones}
\title{Retrieve touchstones a modelling group is responsible for.}
\usage{
montagu_touchstones(modelling_group_id, location = NULL)
}
\arguments{
\item{modelling_group_id}{Id of the modelling group.}

\item{location}{The montagu server to connect to.}
}
\value{
Data frame of touchstone name, description and comment.
}
\description{
Retrieve touchstones a modelling group is responsible for.
}
